I/R: split repl.py into dispatch/cli/server + one-shot cli client
Restructure repl.py into a 45-line argv router that lazily imports one of
two modules, so a one-shot CLI call pays near the Python startup floor
instead of loading the full server:

* repl_cli.py — a stdlib-only one-shot TCP client. `repl.py cli VERB ...`
  connects to a running server, sends a single command, prints the reply,
  exits (the shell analogue of an I/R MCP call, no session startup). Typed
  verbs (init/step/state/fork/...) marshal args into Ir.* ML mirroring the
  Scala IRClient quoting; `raw` sends verbatim; `cli help` lists them.
  Token from --token or $IR_AUTH_TOKEN. Exit 0 ok / 1 ML error / 2 usage.
* repl_srv.py — the full Isabelle/Poly/ML server + console, moved ~verbatim
  from the old repl.py, imported only off the non-cli path.

find-theorems auto-quotes bare term patterns: an unquoted term (e.g.
`sum _ {..<_}`) is an outer-syntax error to find_theorems, so the CLI now
quotes it, matching mcp_server.find_theorems so the two front-ends share
one grammar. A criterion is left literal when already quoted, a name:/simp:
pattern, or a goal keyword (intro/elim/dest/solves); a leading `-` and the
` - ` combinator are preserved.

test_repl.py adds cli_* coverage (typed roundtrip, raw, find-theorems
auto-quote) and a bench_tcp_vs_cli() persistent-TCP-vs-subprocess latency
measurement.
